requirements.txt
altair==5.5.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
fonttools==4.59.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.1
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.3.2
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
protobuf==6.32.0
requests==2.32.5
scikit-learn==1.6.1
imbalanced-learn==0.14.0
scipy==1.16.1
simpy==4.1.1
simuci==1.0.1
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
streamlit==1.48.1
tenacity==9.1.2
threadpoolctl==3.6.0
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
